fix frame rate propagation

+++ src/lib_utils/time.cpp
@@ -55,6 +55,6 @@ void timeInMsToStr(uint64_t timestamp, char buffer[24], const char *msSeparator)
 	uint64_t h = (uint64_t)(p / 3600000);
 	uint8_t m = (uint8_t)(p / 60000 - 60 * h);
 	uint8_t s = (uint8_t)(p / 1000 - 3600 * h - 60 * m);
-	uint16_t u = (uint8_t)(p - 3600000 * h - 60000 * m - 1000 * s);
+	uint16_t u = (uint16_t)(p - 3600000 * h - 60000 * m - 1000 * s);
 	sprintf(buffer, "%02u:%02u:%02u%s%03u", (unsigned)h, (unsigned)m, (unsigned)s, msSeparator, (unsigned)u);
 }
